# Tuning of thresholds for parsing of `BigInt` and `BigDecimal` values