Before:
```
[info] REMEMBER: The numbers below are just data. To gain reusable insights, you need to follow up on
[info] why the numbers are the way they are. Use profilers (see -prof, -lprof), design factorial
[info] experiments, perform baseline and negative tests that provide experimental control, make sure
[info] the benchmarking environment is safe on JVM/OS/HW level, ask for reviews from the domain experts.
[info] Do not assume the numbers tell you what you want them to tell.
[info] Benchmark                           (size)   Mode  Cnt         Score         Error  Units
[info] BigIntBenchmark.readJsoniterScala       10  thrpt    5  25831631.980 ±  266607.498  ops/s
[info] BigIntBenchmark.readJsoniterScala       16  thrpt    5  20301744.606 ± 1301263.280  ops/s
[info] BigIntBenchmark.readJsoniterScala       20  thrpt    5   9437240.012 ±  477855.595  ops/s
[info] BigIntBenchmark.readJsoniterScala       24  thrpt    5   8822253.278 ±  706130.882  ops/s
[info] BigIntBenchmark.readJsoniterScala       28  thrpt    5   8445534.037 ±   57351.107  ops/s
[info] BigIntBenchmark.readJsoniterScala       32  thrpt    5   7529818.595 ±  406724.120  ops/s
[info] BigIntBenchmark.readJsoniterScala       36  thrpt    5   7338671.954 ±  668700.736  ops/s
[info] BigIntBenchmark.readJsoniterScala       40  thrpt    5   6893816.834 ±   40359.106  ops/s
[info] BigIntBenchmark.readJsoniterScala      100  thrpt    5   3358155.335 ±  249144.650  ops/s
[info] BigIntBenchmark.readJsoniterScala      410  thrpt    5    575596.292 ±    8683.137  ops/s
[info] BigIntBenchmark.readJsoniterScala      490  thrpt    5    430442.514 ±    7291.059  ops/s
[info] BigIntBenchmark.readJsoniterScala      570  thrpt    5    322609.882 ±   15505.683  ops/s
[info] BigIntBenchmark.readJsoniterScala      650  thrpt    5    259768.786 ±    1286.222  ops/s
[info] BigIntBenchmark.readJsoniterScala      730  thrpt    5    210222.792 ±    6089.193  ops/s
[info] BigIntBenchmark.readJsoniterScala      810  thrpt    5    206668.555 ±   21438.585  ops/s
[info] BigIntBenchmark.readJsoniterScala     1000  thrpt    5    156917.737 ±    1202.330  ops/s
[info] BigIntBenchmark.readJsoniterScala    10000  thrpt    5      4602.163 ±     155.200  ops/s
[info] BigIntBenchmark.readJsoniterScala   100000  thrpt    5       116.606 ±       6.430  ops/s
[info] BigIntBenchmark.readJsoniterScala  1000000  thrpt    5         3.440 ±       0.013  ops/s
```
After:
```
[info] REMEMBER: The numbers below are just data. To gain reusable insights, you need to follow up on
[info] why the numbers are the way they are. Use profilers (see -prof, -lprof), design factorial
[info] experiments, perform baseline and negative tests that provide experimental control, make sure
[info] the benchmarking environment is safe on JVM/OS/HW level, ask for reviews from the domain experts.
[info] Do not assume the numbers tell you what you want them to tell.
[info] Benchmark                           (size)   Mode  Cnt         Score         Error  Units
[info] BigIntBenchmark.readJsoniterScala       10  thrpt    5  29846691.320 ± 1393660.311  ops/s
[info] BigIntBenchmark.readJsoniterScala       16  thrpt    5  24080437.254 ±  210922.894  ops/s
[info] BigIntBenchmark.readJsoniterScala       20  thrpt    5  11380702.462 ±   97416.979  ops/s
[info] BigIntBenchmark.readJsoniterScala       24  thrpt    5  11122764.542 ±  663579.974  ops/s
[info] BigIntBenchmark.readJsoniterScala       28  thrpt    5  10573171.658 ±  734507.744  ops/s
[info] BigIntBenchmark.readJsoniterScala       32  thrpt    5   9693864.248 ±  503946.093  ops/s
[info] BigIntBenchmark.readJsoniterScala       36  thrpt    5   9468383.347 ±  546329.091  ops/s
[info] BigIntBenchmark.readJsoniterScala       40  thrpt    5   6869447.924 ±  432367.517  ops/s
[info] BigIntBenchmark.readJsoniterScala      100  thrpt    5   3507631.321 ±   28301.497  ops/s
[info] BigIntBenchmark.readJsoniterScala      410  thrpt    5    564767.757 ±   40662.410  ops/s
[info] BigIntBenchmark.readJsoniterScala      490  thrpt    5    513738.164 ±   31675.540  ops/s
[info] BigIntBenchmark.readJsoniterScala      570  thrpt    5    425258.466 ±    2990.584  ops/s
[info] BigIntBenchmark.readJsoniterScala      650  thrpt    5    291495.825 ±   16908.786  ops/s
[info] BigIntBenchmark.readJsoniterScala      730  thrpt    5    250610.030 ±   12205.738  ops/s
[info] BigIntBenchmark.readJsoniterScala      810  thrpt    5    207094.461 ±   19163.991  ops/s
[info] BigIntBenchmark.readJsoniterScala     1000  thrpt    5    183021.548 ±    6219.874  ops/s
[info] BigIntBenchmark.readJsoniterScala    10000  thrpt    5      4685.921 ±      24.971  ops/s
[info] BigIntBenchmark.readJsoniterScala   100000  thrpt    5       115.813 ±       5.190  ops/s
[info] BigIntBenchmark.readJsoniterScala  1000000  thrpt    5         3.475 ±       0.032  ops/s
```